PR#7506: Range Locking: Implementation: fix issues detected by CI
Continuation of
 PR#7506: Range Locking: Implementation: fix issues detected by CI part 1
part 2

+++ utilities/transactions/lock/range/range_tree/locktree/wfg.cc
@@ -131,7 +131,7 @@ bool wfg::cycle_exists_from_txnid(TXNID txnid,
 void wfg::apply_nodes(int (*fn)(TXNID id, void *extra), void *extra) {
   int r = 0;
   uint32_t n_nodes = m_nodes.size();
-  for (size_t i = 0; i < n_nodes && r == 0; i++) {
+  for (uint32_t i = 0; i < n_nodes && r == 0; i++) {
     node *n;
     r = m_nodes.fetch(i, &n);
     invariant_zero(r);
